Suppression de JEvent
 * Qui étend JPanel, mais n'est jamais utilisé en tant que tel.
 * Qui ne définit qu'une seule méthode auto-suffisante.
