+++ swing/src/main/java/io/github/algorys/agshmne/game/south/JDescGame.java
@@ -1,17 +1,29 @@
 package io.github.algorys.agshmne.game.south;
 
 import java.awt.Dimension;
+import java.awt.GridBagConstraints;
+import java.awt.GridBagLayout;
+import java.awt.Insets;
+import java.awt.event.ActionEvent;
+import java.awt.event.ActionListener;
 
 import javax.swing.BoxLayout;
+import javax.swing.JButton;
+import javax.swing.JDialog;
+import javax.swing.JFrame;
+import javax.swing.JLabel;
 import javax.swing.JPanel;
+import javax.swing.SwingUtilities;
 
+import io.github.algorys.agshmne.character.opponent.beast.Beast;
+import io.github.algorys.agshmne.character.opponent.beast.BeastFactory;
 import io.github.algorys.agshmne.character.player.Player;
+import io.github.algorys.agshmne.fight.Fight;
 
 @SuppressWarnings("serial")
 public class JDescGame extends JPanel {
-	JTitleHistory titleHistory;
-	JCivilized civilized;
-	JEvent event;
+	private JTitleHistory titleHistory;
+	private JCivilized civilized;
 	
 	public JDescGame(Player pj){
 		this.setPreferredSize(new Dimension(710, 100));
@@ -27,8 +39,7 @@ public class JDescGame extends JPanel {
 		this.civilized.setCivilized(pj.getTile().isCivilized());
 		
 		// TODO Danger aux alentour (Rouge)
-		this.event = new JEvent(pj);
-		this.event.updateEvent(pj.getTile().isDanger());
+		this.updateEvent(pj);
 		
 		// TODO Actions Personnage**this.add(titleHistory);
 		
@@ -40,6 +51,86 @@ public class JDescGame extends JPanel {
 	public void updateLabels(Player pj) {
 		titleHistory.setTitle("Region : " + pj.getTile().getDesc());
 		civilized.setCivilized(pj.getTile().isCivilized());
-		event.updateEvent(pj.getTile().isDanger());
+		this.updateEvent(pj);
+	}
+	
+	public void updateEvent(Player pj) {
+		boolean danger = pj.getTile().isDanger();
+		if (danger) {
+			Beast wolf = new BeastFactory().createBeast();
+			final Fight fight = new Fight(pj, wolf);
+			initJFight(pj, wolf, fight);
+		}
+	}
+
+	private void initJFight(Player pj, Beast wolf, final Fight fight) {
+		JFrame topFrame = (JFrame) SwingUtilities.getWindowAncestor(this);
+		JDialog jDialogEvent = new JDialog(topFrame, "Vous êtes attaqué !", true);
+		jDialogEvent.setSize(600, 300);
+		jDialogEvent.setLocationRelativeTo(topFrame);
+
+		JPanel panEvent = new JPanel();
+		panEvent.setLayout(new GridBagLayout());
+		GridBagConstraints gbcEvent = new GridBagConstraints();
+		gbcEvent.insets = new Insets(5, 5, 5, 5);
+
+		gbcEvent.gridy = 0;
+		gbcEvent.gridheight = 1;
+		gbcEvent.gridx = 0;
+		gbcEvent.gridwidth = 4;
+		gbcEvent.anchor = GridBagConstraints.CENTER;
+		gbcEvent.fill = GridBagConstraints.NONE;
+		panEvent.add(new JLabel("COMBAT"), gbcEvent);
+
+		// PLAYER
+		gbcEvent.gridy = 1;
+		gbcEvent.gridheight = 1;
+		gbcEvent.gridx = 0;
+		gbcEvent.gridwidth = 2;
+		gbcEvent.anchor = GridBagConstraints.WEST;
+		gbcEvent.fill = GridBagConstraints.NONE;
+		panEvent.add(new JCharacter(pj), gbcEvent);
+
+		// MONSTER
+		gbcEvent.gridy = 1;
+		gbcEvent.gridheight = 1;
+		gbcEvent.gridx = 2;
+		gbcEvent.gridwidth = 2;
+		gbcEvent.anchor = GridBagConstraints.EAST;
+		gbcEvent.fill = GridBagConstraints.NONE;
+		panEvent.add(new JCharacter(wolf), gbcEvent);
+
+		JButton jbAttaq = new JButton("Attaquer");
+		gbcEvent.gridy = 5;
+		gbcEvent.gridheight = 1;
+		gbcEvent.gridx = 2;
+		gbcEvent.gridwidth = 2;
+		gbcEvent.anchor = GridBagConstraints.EAST;
+		gbcEvent.fill = GridBagConstraints.NONE;
+		panEvent.add(jbAttaq, gbcEvent);
+
+		final JLabel outputAttaq = new JLabel("Init combat...");
+
+		jbAttaq.addActionListener(new ActionListener() {
+
+			@Override
+			public void actionPerformed(ActionEvent e) {
+				fight.newRound();
+				fight.pjAttack();
+				fight.mobAttack();
+				outputAttaq.setText("Vous attaquez !");
+			}
+		});
+
+		gbcEvent.gridy = 7;
+		gbcEvent.gridheight = 1;
+		gbcEvent.gridx = 1;
+		gbcEvent.gridwidth = 4;
+		gbcEvent.anchor = GridBagConstraints.CENTER;
+		gbcEvent.fill = GridBagConstraints.NONE;
+		panEvent.add(outputAttaq, gbcEvent);
+
+		jDialogEvent.add(panEvent);
+		jDialogEvent.setVisible(true);
 	}
 }
